V0.10 (#13)
* Final rewrite, I promise. This will be implemented on top of Slacktor as a backend. The idea is to enable much, much higher performance while simplifying code significantly.

* Removed everything that will be changed by the rewrite.

* Working on identifiers and adding actors.

* Working on actor traits.

* Note to future self: The ActorWrapper can be cloned into an owned type as it contains an Arc.
This allows us to essentially bypass the RwLock on the Slacktor instance, meaning messages can still be sent as actors are added and removed. All that will be blocked by mutable operations on the RwLock are actor retrievals.

* Basic actor retrieval done.

* Hmmm. Need a way to enforce serialize on foreign actors only...

* Lots of work done.
It is now possible to utilize messages that both implement and dont implement serialization functionality.
Obviously, messages that do not implement serialization functionality can not be used on foreign systems as long as the 'serde' feature is enabled.
The 'serde' feature is now independent of the 'foreign' feature. Both must be enabled to allow delegates to serialize messages.

* Starting on updating docs. I am thinking that after publishing version 0.10, I am going to test fluxion with a few basic programs, write a few more tutorials and documentation pages, and publish version 1.0.0

* Working on docs, need to do actor contexts.

* Actor contexts done.

* Nicely commented the example.
I am going to get rid of the current documentation, as it is pretty overkill.
Fluxion has lost *a lot* of weight this update.
Some may come back in the future, however I think that Fluxion is much better when slimmed down this way, and doesn't get in the way as much.
Fluxion's goal has always been to be an actor system with one unique feature, that doesn't require you to structure your code a specific way, but also is extremely transparent, meaning no hidden async executors or tasks running in the background.

* hmm. For some reason, unable to compile completely no-std.

* Make sure core is built.

* New platform for generating site.

* Working on changelog and new documentation.
Now using MDbook and oranda for a nice site.

* Basic benchmark, add inline to a few functions that may benefit.

* Add disclaimer to benchmark

* Writing Foreward and new readme.

* I am seeing an issue with delegates... How do you determine which type to deserialize to?

* Foreign proxies will be required, just not on the fluxion system.

* Fixed some minor issues with serialization.

* Moving to workspace-based file structure, as a proc-macro crate will be required for message ids.
Note: For ensure_no_std to compile, nightly with -Zbuild-std is required.

* Message proc macro done. Now need a mechanism to automatically create a register of all message types initiated...

* I think this will work!

* Working on foreign delegate example.
Added ability to get an actor's id from a local ref.

* using channels is getting cluttered, I might actually just use slacktor for the delegates.

* Delegate is implemented for Arcs of delegates automatically.

Bam. Baking message ID support into fluxion allows delegates to serialize and deserialize messages to and from foreign systems.
The system still requires registering messages for actors, but that is handled by the delegate and not by Fluxion itself.

* Re-organizing project a little, putting oranda docs back into fluxion folder.

* Actually, lets put that back.

* Working on docs and preparing for release.

* Remove old docs.

* I am going to say that is enough documentation for now. I think we are clear to merge into master.

* Bump fluxion version, fix clippy lints.

* Update changelog.

* Add oranda github pages action.

* Update web.yml.